Added Algorithm to Combat Stack Overflow
This Algorthim is built to fight stack overflow when automatically revealing boxes. Stack Overflow used to happen if there where no bombs on a over 60*50 map. Now we seem to be safe from stack overflow with a 100*100 map with no bombs. This has not been tested or thouroughly proven. It also could be optimized, but it should insure that a 50*50 map is no chance of stack overflow. Ask Brandon if you want to know how the trick works.