Added Number of Bombs and Safe Spaces
I added a label at the top of the game form that displays the number of bombs left unmarked. I also added a label at the top of the game form that displays the number of safe spaces left unmarked or unturned. Both of these values need to be set to zero when the game is over.